Fetching programs moved to root component
