+++ client/student/components/Programs/Activity/ContentContainer.vue
@@ -13,7 +13,7 @@
 
 <script>
 import api from '@/student/api/content';
-import CircularProgress from '../common/CircularProgress';
+import CircularProgress from '@/student/components/common/CircularProgress';
 import TailorTeachingElements from 'tailor-teaching-elements';
 
 export default {
