+++ client/student/components/Programs/Courseware/ActivityCard.vue
